Add a Console application that can read a database
using the same context, even though it's not the same ephemeral database
that the Web project will use (it's an in-process version of SQL Server
Express).

I had to add a reference to EntityFramework (why didn't that come
through as a transitive dependency?), and I had to copy the connection
string from `<web>/web.config` to `<cli>/app.config` to get it to work.